Moved solid interface from index to utils, with @yustarandomname
