+++ app/svelte/src/lib/stores/userStore.ts
@@ -11,6 +11,8 @@ import {
 interface UserStore {
 	loading: boolean;
 	userSession: ISessionInfo;
+	webID: string;
+	podUrl: string;
 }
 
 /***
@@ -19,26 +21,33 @@ interface UserStore {
 function createUserStore() {
 	const { subscribe, set } = writable<UserStore>({
 		loading: true,
-		userSession: getDefaultSession().info
+		userSession: getDefaultSession().info,
+		webID: '',
+		podUrl: ''
 	});
 
 	return {
 		subscribe,
 		set,
-		init: async () => {
+		init: async (webID?: string, podUrl?: string) => {
 			// Check if there is a session in the browser
 			await handleIncomingRedirect({
 				restorePreviousSession: true
 			});
 
+			webID = webID || localStorage.getItem('webID') || 'localhost:3000';
+			podUrl = podUrl || localStorage.getItem('podUrl') || 'johndoe';
+
 			set({
 				loading: false,
-				userSession: getDefaultSession().info
+				userSession: getDefaultSession().info,
+				webID,
+				podUrl
 			});
 		},
-		signIn: async (podUrl: string, webId: string, interfaceUrl: string) => {
+		signIn: async (podUrl: string, webID: string, interfaceUrl: string) => {
 			if (!podUrl) throw new Error('Please enter a pod URL');
-			if (!webId) throw new Error('Please enter a web ID');
+			if (!webID) throw new Error('Please enter a web ID');
 
 			if (podUrl.indexOf('//') < 0) {
 				podUrl = 'http://' + podUrl;
@@ -46,7 +55,7 @@ function createUserStore() {
 
 			// Set podUrl and webId for next sessions
 			localStorage.setItem('podUrl', podUrl);
-			localStorage.setItem('webID', webId);
+			localStorage.setItem('webID', webID);
 
 			// If something goes wrong, the error message will be displayed in the UI
 			try {
@@ -61,7 +70,9 @@ function createUserStore() {
 
 				set({
 					loading: false,
-					userSession: getDefaultSession().info
+					userSession: getDefaultSession().info,
+					webID,
+					podUrl
 				});
 			} catch (e) {
 				throw new Error(getErrorMessage(e).message);
